Add API endpoints to allow modification of surveys and diaries.
This is meant to allow clean-up of already-entered surveys and diaries
in exceptional cases and only if users explicitly request it.

The new endpoints are protected by HTTP basic auth, plus requests need
to come from well-known IP addresses.  Both the allowed IP addresses
and the authorized username/password pairs are set in the
configuration file.